fix(config): use rad node root url as http endpoint config default
Previouly the default http endpoint config value would include the
`/api/v1` path. That should be abstracted as it is an application
implementation detail. This commit changes the default config value to
not include it.
Additionally applies some light refactoring and clean-up to api-
related code.


#26
