+++ package.json
@@ -245,7 +245,7 @@
           "radicle.advanced.httpApiEndpoint": {
             "scope": "machine-overridable",
             "type": "string",
-            "default": "http://127.0.0.1:8080/api/v1",
+            "default": "http://127.0.0.1:8080",
             "markdownDescription": "Specifies the endpoint to the root of the Radicle HTTP API connected to a seed node. This is commonly served via `radicle-httpd` and can be any Fully-Qualified Domain Name (FQDN) i.e. $protocol://$host:$port/$path ."
           }
         }
